removes unnecessary authenticateUser module. Refactors out <AuthProvider/> which removes the need of await in tests for waiting auth fetch requests to be completed
