+++ src/components/Home.test.js
@@ -1,14 +1,11 @@
 import React from 'react';
 import { Home } from './Home';
-import { renderWithRouter, authenticateUser, waitForElement } from 'test-utils';
+import { renderWithRouter } from 'test-utils';
 
 it('renders <Home/> without crashing', async () => {
-  const reset = authenticateUser();
-
   const { queryByText } = renderWithRouter(<Home />, { router: '/' });
 
-  //Wait for Home page to be loaded
-  await waitForElement(() => queryByText(/your playlists/i));
+  const PlaylistHeading = queryByText(/your playlists/i);
 
-  reset();
+  expect(PlaylistHeading).toBeTruthy();
 });
