+++ src/components/PlaylistEditor.test.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import { PlaylistEditor } from './PlaylistEditor';
-import { renderWithRouter, authenticateUser } from 'test-utils';
+import { renderWithRouter } from 'test-utils';
 
 it('renders <PlaylistEditor/> without crashing', () => {
-  authenticateUser();
   renderWithRouter(<PlaylistEditor />, {
     router: '/playlist-editor',
   });
